[Pal] Drop support for uri==NULL in DkProcessCreate
This feature doesn't seem to make much sense. Also, it was already
broken: Pal/regression/Process3.c assumed uri==NULL means spawning the
process without an executable, but PAL code interpreted this as "spawn
caller's executable again". The documentation doesn't mention this
special case, so it's hard to say which semantics was originally
intended.

+++ Pal/src/host/Linux-SGX/enclave_ocalls.h
@@ -71,7 +71,7 @@ int ocall_resume_thread(void* tcs);
 
 int ocall_clone_thread(void);
 
-int ocall_create_process(const char* uri, int nargs, const char** args, int* stream_fd,
+int ocall_create_process(const char* uri, size_t nargs, const char** args, int* stream_fd,
                          unsigned int* pid);
 
 int ocall_futex(uint32_t* uaddr, int op, int val, int64_t timeout_us);
